+++ test/src/viriontest/sofe/nbtstreams/Test.php
@@ -1,0 +1,26 @@
+<?php
+
+/*
+ *
+ * nbtstreams
+ *
+ * Copyright (C) 2017 SOFe
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ */
+
+declare(strict_types=1);
+
+namespace viriontest\sofe\nbtstreams;
+
+use pocketmine\plugin\PluginBase;
+
+class Test extends PluginBase{
+	public function onEnable(){
+		// TODO Write tests
+	}
+}